dom/xslt/tests/mochitest/test_bug440974.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=440974
-->
<head>
  <title>Test for Bug 440974</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=440974">Mozilla Bug 440974</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 440974 **/

function isTxResult(node)
{
  return node.namespaceURI == "http://www.mozilla.org/TransforMiix" &&
         node.localName == "result";
}

var xmldoc, xsltdoc;
[ xmldoc, xsltdoc ] = [ new DOMParser().parseFromString(xml, "text/xml") for (xml of [ 

    '<items><item><id>1</id></item><item><id>2</id></item><item><id>3</id></item></items>' ,

    '<xsl:stylesheet version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform">\n\
        <xsl:output method="xml" />\n\
        <xsl:template match="item"><foo id="{id}"/></xsl:template>\n\
    </xsl:stylesheet>'

]) ];
var processor = new XSLTProcessor;
processor.importStylesheet(xsltdoc);
var result = processor.transformToDocument(xmldoc);
var resultElements = Array.prototype.filter.call(result.getElementsByTagName('*'), isTxResult);
is(resultElements.length, 1, "there should be only one 'transformiix:result' element"); 
is(resultElements[0], result.documentElement, "the 'transformiix:result' element should be the document element"); 

</script>
</pre>
</body>
</html>
